+++ tests/testthat/test-PKNCA.options.R
@@ -58,7 +58,6 @@ test_that("PKNCA.options", {
                list(adj.r.squared.factor=0.0001,
                     max.missing=0.5,
                     auc.method="lin up/log down",
-                    conc_above=NA_real_,
                     conc.na="drop",
                     conc.blq=list(
                       first="keep",