+++ tests/testthat/test-time.above.R
@@ -1,13 +1,15 @@
-context("time_above")
-
 test_that("time_above expected errors", {
   expect_error(
-    pk.calc.time_above(time="X", conc_above="X"),
-    regexp='argument "method" is missing, with no default'
+    pk.calc.time_above(conc=c(1, 1), time=c(1, 2), conc_above="X", method="linear"),
+    regexp='conc_above must be numeric'
   )
   expect_error(
-    pk.calc.time_above(time="X", conc_above="X", method="linear"),
-    regexp='conc_above must be numeric'
+    pk.calc.time_above(conc=c(1, 1), time=c(1, 2), conc_above=1:2, method="linear"),
+    regexp='conc_above must be a scalar'
+  )
+  expect_error(
+    pk.calc.time_above(conc=c(1, 1), time=c(1, 2), conc_above=NA, method="linear"),
+    regexp='conc_above must not be NA'
   )
   expect_error(
     pk.calc.time_above(time="X", conc_above=5, method="linear"),
@@ -63,5 +65,13 @@ test_that("time_above linear", {
     pk.calc.time_above(conc=c(6, NA, 5), time=1:3, conc_above=5, method="linear"),
     2,
     info="NA is ignored"
+  )
+})
+
+test_that("time_above with 'lin up/log down'", {
+  skip("not yet implemented")
+  expect_equal(
+    pk.calc.time_above(conc=c(6, 4, 5), time=1:3, conc_above=5, method='lin up/log down'),
+    0.5
   )
 })